feat(memos): open in-stream memo previews in a modal/drawer
Clicking an in-stream memo preview card now opens a responsive preview
dialog — a centered modal on desktop, a bottom drawer on mobile — instead
of navigating away to the memory explorer. The dialog renders the full
memo (title, abstract, key points, provenance, source messages) and links
through to the explorer for the full surface.

The memo detail renderer is extracted from the memory explorer page into a
shared module so the dialog and the page render a memo identically, and the
dialog shares the memo-detail query cache with the embed card's resolver so
opening an already-resolved card is instant.
